Moved to inventory package
