+++ internal/v1_12_R1/src/main/java/net/flawe/offlinemanager/api/util/v1_12_R1/inventory/OfflineEnderChest.java
@@ -1,11 +1,16 @@
-package net.flawe.offlinemanager.api.util.v1_16_R1;
+package net.flawe.offlinemanager.api.util.v1_12_R1.inventory;
 
 import net.flawe.offlinemanager.api.inventory.AbstractInventory;
 import net.flawe.offlinemanager.api.inventory.IEnderChest;
-import net.minecraft.server.v1_16_R1.*;
-import org.bukkit.craftbukkit.v1_16_R1.inventory.CraftInventory;
-import org.bukkit.craftbukkit.v1_16_R1.inventory.CraftItemStack;
+import net.minecraft.server.v1_12_R1.IInventory;
+import net.minecraft.server.v1_12_R1.ItemStack;
+import net.minecraft.server.v1_12_R1.NBTTagCompound;
+import net.minecraft.server.v1_12_R1.NBTTagList;
+import org.bukkit.craftbukkit.v1_12_R1.inventory.CraftInventory;
+import org.bukkit.craftbukkit.v1_12_R1.inventory.CraftItemStack;
 import org.bukkit.inventory.Inventory;
+
+import java.util.HashMap;
 
 public class OfflineEnderChest extends AbstractInventory implements IEnderChest {
 
@@ -18,12 +23,11 @@ public class OfflineEnderChest extends AbstractInventory implements IEnderChest 
         NBTTagList list = (NBTTagList) tag.get("EnderItems");
         if (list == null)
             throw new NullPointerException("EnderItems cannot be null!");
-        for (NBTBase base : list) {
-            if (!(base instanceof NBTTagCompound)) continue;
-            NBTTagCompound item = (NBTTagCompound) base;
+        for (int i = 0; i < list.size(); i++) {
+            NBTTagCompound item = list.get(i);
             int slot = item.getByte("Slot") & 0xFF;
             if (slot < inventory.getSize())
-                inventory.setItem(slot, CraftItemStack.asBukkitCopy(ItemStack.a(item)));
+                inventory.setItem(slot, CraftItemStack.asBukkitCopy(new ItemStack(item)));
         }
         this.tag = tag;
     }
@@ -47,5 +51,41 @@ public class OfflineEnderChest extends AbstractInventory implements IEnderChest 
     @Override
     public Inventory getEnderChest() {
         return this;
+    }
+
+    @Override
+    public String getName() {
+        return inventory.getName();
+    }
+
+    @Override
+    public boolean contains(int i) {
+        return inventory.contains(i);
+    }
+
+    @Override
+    public boolean contains(int i, int i1) {
+        return inventory.contains(i, i1);
+    }
+
+    @Override
+    public HashMap<Integer, ? extends org.bukkit.inventory.ItemStack> all(int i) {
+        return inventory.all(i);
+    }
+
+    @Override
+    public int first(int i) {
+        return inventory.first(i);
+    }
+
+    @Override
+    public void remove(int i) {
+        inventory.remove(i);
+        update();
+    }
+
+    @Override
+    public String getTitle() {
+        return inventory.getTitle();
     }
 }
