+++ internal/v1_16_R1/src/main/java/net/flawe/offlinemanager/api/util/v1_16_R1/inventory/OfflineInventory.java
@@ -1,4 +1,4 @@
-package net.flawe.offlinemanager.api.util.v1_16_R1;
+package net.flawe.offlinemanager.api.util.v1_16_R1.inventory;
 
 import net.flawe.offlinemanager.api.inventory.IInventory;
 import org.bukkit.Bukkit;
