+++ internal/v1_16_R3/src/main/java/net/flawe/offlinemanager/api/util/v1_16_R3/inventory/PlayerInventory.java
@@ -8,13 +8,11 @@ import org.bukkit.craftbukkit.v1_16_R3.inventory.CraftInventoryPlayer;
 
 public class PlayerInventory extends AbstractPlayerInventory {
 
-    private final org.bukkit.inventory.PlayerInventory inventory;
     private final NBTTagCompound tag;
 
     public PlayerInventory(org.bukkit.inventory.PlayerInventory inventory, NBTTagCompound tag) {
         super(inventory);
         this.tag = tag;
-        this.inventory = inventory;
     }
 
     @Override
